requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# LLM and AI dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0

# Database dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0
redis>=5.0.0

# Vector store dependencies
pinecone-client>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Local LLM support
ctransformers>=0.2.0
sentence-transformers>=2.2.0 